Final Update (99%)
Fixed neighbors() in Board to ensure a replicate of the original wasn't
returned. 

Made SearchNode Comparable. Optimized the constructor by making sure
duplicate nodes are not added to the Priority Queue. 

Only issue is with the iterator. One of tests wants the goal node first,
while the other wants the initial node to be first. Am not sure how to
accommodate both.